Move targzipwriter to common module
